multithreading support for multi-step measurements

+++ core/src/main/java/com/yahoo/ycsb/measurements/Measurements.java
@@ -145,7 +145,7 @@ public class Measurements
         }
     }
 
-    public void measureStepTimeout(String operation, double timeout) {
+    public void measureStepTimeout(String operation, int timeout) {
         try {
             msData.get(operation).measureTimeout(timeout);
         }
